✨ Loan Accept Decimal value and set back to active

+++ src/components/Admin/Loan/loan.js
@@ -86,7 +86,7 @@ const Loan = ({ match }) => {
       return;
     }
 
-    setAmount(parseInt(val.replace(/\D/, ''), 10))
+    setAmount(parseFloat(val.replace(/\D/, ''), 10))
   }
 
   const handlePenalties = total => {
@@ -316,7 +316,7 @@ const Loan = ({ match }) => {
             </div>
             <div className="info">
               <p className="label">Loan Amount:</p>
-              <p className="value">&#8369;{monify(parseInt(amount))}</p>
+              <p className="value">&#8369;{monify(parseFloat(amount))}</p>
             </div>
             <div className="info">
               <p className="label">Loan Terms:</p>
@@ -363,19 +363,19 @@ const Loan = ({ match }) => {
                   <div className="computation-container">
                     <div className="computations">
                       <p className="label">Principal:</p>
-                      <p className="value">{monify(parseInt(amount))}</p>
+                      <p className="value">{monify(parseFloat(amount))}</p>
                     </div>
                     <div className="computations">
                       <p className="label">Finance Charge &#40; {financeCharge}% &#41;:</p>
-                      <p className="value">&#40;{(parseInt(amount) * (financeCharge / 100)).toFixed(2)}&#41;</p>
+                      <p className="value">&#40;{(parseFloat(amount) * (financeCharge / 100)).toFixed(2)}&#41;</p>
                     </div>
                     <div className="computations">
                       <p className="label">Processing Fee &#40; 2% &#41;:</p>
-                      <p className="value">&#40;{(parseInt(amount) * (2 / 100)).toFixed(2)}&#41;</p>
+                      <p className="value">&#40;{(parseFloat(amount) * (2 / 100)).toFixed(2)}&#41;</p>
                     </div>
                     <div className="computations">
                       <p className="label">Service Fee &#40; {serviceFee}% &#41;:</p>
-                      <p className="value">&#40;{(parseInt(amount) * (serviceFee / 100)).toFixed(2)}&#41;</p>
+                      <p className="value">&#40;{(parseFloat(amount) * (serviceFee / 100)).toFixed(2)}&#41;</p>
                     </div>
                     <div className="computations loan-proceeds">
                       <p className="label">Loan Proceeds:</p>
@@ -395,12 +395,12 @@ const Loan = ({ match }) => {
                   <div className="computation-container">
                     <div className="computations">
                       <p className="label">Principal:</p>
-                      <p className="value">&#8369;{monify(parseInt(loanResult.amount))}</p>
+                      <p className="value">&#8369;{monify(parseFloat(loanResult.amount))}</p>
                     </div>
                     <div className="computations">
                       <p className="label">Loan Proceeds:</p>
                       <p className="value">&#8369;{
-                        parseInt(
+                        parseFloat(
                           loanResult.amount
                           - (loanResult.amount * (2 / 100))
                           - (loanResult.amount * (loanResult.serviceFee / 100))
@@ -411,7 +411,7 @@ const Loan = ({ match }) => {
                     <div className="computations">
                       <p className="label">Total Charges:</p>
                       <p className="value">&#8369;{
-                        parseInt(
+                        parseFloat(
                           loanProceedsValue +
                           (amount * (2 / 100))
                           + (amount * (serviceFee / 100))
@@ -422,27 +422,27 @@ const Loan = ({ match }) => {
                     </div>
                     <div className="computations sub-comp">
                       <p className="label">Finance Charge &#40; {financeCharge}% &#41;:</p>
-                      <p className="value">{parseInt(amount * (financeCharge / 100)).toFixed(2)}</p>
+                      <p className="value">{parseFloat(amount * (financeCharge / 100)).toFixed(2)}</p>
                     </div>
                     <div className="computations sub-comp">
                       <p className="label">Processing Fee &#40; 2% &#41;:</p>
-                      <p className="value">{parseInt(amount * (2 / 100)).toFixed(2)}</p>
+                      <p className="value">{parseFloat(amount * (2 / 100)).toFixed(2)}</p>
                     </div>
                     <div className="computations sub-comp">
                       <p className="label">Service Fee &#40; {serviceFee}% &#41;:</p>
-                      <p className="value">{parseInt(amount * (serviceFee / 100)).toFixed(2)}</p>
+                      <p className="value">{parseFloat(amount * (serviceFee / 100)).toFixed(2)}</p>
                     </div>
                     <div className="computations sub-comp">
                       <p className="label">Penalty:</p>
-                      <p className="value">{parseInt(penaltyCharge).toFixed(2)}</p>
+                      <p className="value">{parseFloat(penaltyCharge).toFixed(2)}</p>
                     </div>
                     <div className="computations after-charges">
                       <p className="label">Total Payment:</p>
-                      <p className="value">&#8369;{monify(parseInt(totalPayment))}</p>
+                      <p className="value">&#8369;{monify(parseFloat(totalPayment))}</p>
                     </div>
                     <div className="computations">
                       <p className="label">Balance:</p>
-                      <p className="value">&#8369;{monify(parseInt(totalBalance))}</p>
+                      <p className="value">&#8369;{monify(parseFloat(totalBalance))}</p>
                     </div>
                   </div>
                 </div>
